MF Edits
Hi, Roel!

Thank you for this post! It was an easy read and written well. At the line level, I didn't make too many adjustements. I did have a few TODOs I left for you, however, as there were some points of clarification needed.

There's a couple of things I'd encourage you to do moving foward, which are also addressed with my TODOs.

You could do a better job linking your ideas to each other. Through the post, you tend to mention an idea and then give code and then say another idea. It isn't explicitly clear that the code you are giving is directly related to the idea that you just gave or the idea after the code you are giving is related to the code that just came before.

You'll usually want to follow some kind of framing, even if it is lose:
- Here's an idea
- I can demonstrate this idea with some code. I'll do that now
- Gives code
- As you can see, x and y commentary about the idea or code.

It can be concise and loose, but some kind of framing is generally encouraged. I've left TODOs to ask you to do this.

Another thing I'd encourage you to avoid in service of clarity is starting paragraphs where the subject is a pronoun. I know you know what  pronoun is but I'll review anyway -- a word that stands in for a noun, usually "she / him / it / this / that / etc." In your case, you have a couple of paragraphs that start with "This." The issue is that I don't know if "This" is referring to the code above or the idea you discussed in the paragraph above the code or the problem with the site you are fixing or the action you took or... I can't tell if the "this" is a direct analysis of what you just discussed or an overall commentary on a larger idea, like SQL. I'd recommend always starting a new paragraph with the noun, and not a pronoun.

A small comment: Per the style guide, Boolean is capped.

One last comment about content inclusion: So I think engineering spent time engineering these search functions because it would solve some kind of problem -- you weren't just noodling to noodle and see what you could do. You gesture to the problem that you want solved but don't really get at it that explicitly. You'll see one of the TODOs I left you asked you to flesh out a little bit more what the issue is. What was the reader EXPECTING to see in their search, what did they see, what did you do to resolve it and now what does the resolved result look like? Again, I think you gesture toward it -- I can gather that's what you're talking about -- but more specificity would be good, otherwise the piece reads more like an experiential account of what you did instead of a helpful guide for readers to self-idenify if your engineering issue is something they might also run into.

Feel free, as always, to disregard any of my TODOs; they are just suggestions.

Thank you, again, and let me know if you want me to proofread any content you add if you add any.

Best,
Manda